requirements.txt
# DO NOT include azure-functions-worker in this file
# The Python Worker is managed by Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

azure-functions
azureml-core
scikit-learn==1.3.0
pandas==2.0.3
numpy==1.25.2
# flask
# flask-cors
